update brick.c / beginning of isLit on test.c

+++ renderer/ellipse.c
@@ -65,7 +65,7 @@ Point contactEllipseWithLine(Ellipse E, Line L) {
 */
 double * encodeEllipse(Ellipse E){
     double *ellipse;
-    ellipse = malloc(6 * sizeof(int));
+    ellipse = malloc(6 * sizeof(double));
     ellipse[0] = E.a;
     ellipse[1] = E.b;
     ellipse[2] = E.c;
